requirements.txt
flask==3.0.0
requests==2.31.0
openai>=1.50.0
chromadb==0.4.22
numpy<2.0
python-dotenv==1.0.0
